+++ NEWS.md
@@ -2,9 +2,10 @@
 
 ## Other changes
 
+- Slight speed-up of `kernelshap()` and `permshap()` for single-output predictions.
+- Slight speed-up of `kernelshap()` and `permshap()` for factor-valued predictions.
 - Slight speed-up of `permshap()` by saving calculations for the two special permutations of all 0 and all 1.
 - Consequently, the `m_exact` component in the output is reduced by 2.
-- Slight speed-up of `kernelshap()` and `permshap()` for single-output predictions.
 
 # kernelshap 0.4.0